cmake: add CMake option BUILD_OROGEN (defaults to OFF)
orogen dependency utilrb depends on metaruby (https://github.com/rock-core/tools-metaruby),
a package that is not available from default package repostories of common Linux
distributions. It needs to be installed as a Ruby gem or from source.

Until a unified way of bootstrapping the toolchain with or without Ruby and optional
packages I therefore suggest to disable orogen for CMake-based builds by default.